Update cursor wrap-around behaviour
Now that you can get betweent the top area and bottom area using the
up and down keys, there is no need to have the crisscross behaviour
when moving off the side of the screen.

Now, when navigating off the side of the screen, the cursor doesn't jump
between the upper and lower areas.