Improve animation when cancelling a card move
To animate to the final position, on top of another card, the animation
added a +1 to the destination. However, this doesn't properly for cards
returning to the bottom of a stack.

Instead, when picking a card just store the depth so that it can be used
in the animation if the move is cancelled.