Simplify slide animation code, fix frame with missing cursor
Previously, at the end of a cancel animation, there would be one frame
where the cursor would not be drawn. This change fixes that issue.

Note that, there are still some cases where a card may be missing from
a single frame at the start of the animation, likely caused by trying to
do too much during vblank.